# src/main/resources/openapi/disk-sharing.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
paths:
  /sharing/disks/take:
    put:
      tags:
        - sharing-controller
      operationId: takeDisk
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TakeDiskRequestMo'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /sharing/disks/return:
    put:
      tags:
        - sharing-controller
      operationId: returnDisk
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnDiskRequestMo'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    TakeDiskRequestMo:
      type: object
      properties:
        diskId:
          type: integer
          format: int64
    ReturnDiskRequestMo:
      type: object
      properties:
        diskId:
          type: integer
          format: int64
...